pyproject.toml
[project]
name = "tut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.29.5",
    "langgraph>=0.5.1",
    "python-dotenv>=1.1.1",
    "langchain[groq,openai]>=0.3.26",
    "langchain-openai>=0.1.6",
]